pyproject.toml
[build-system]
requires = ["setuptools>=80", "wheel"]
build-backend = "setuptools.build_meta"
# requires = ["hatchling"]
# build-backend = "hatchling.build"

[project]
name = "machineconfig"
version = "3.0"
description = "Dotfiles management package"
readme = "README.md"
license = {text = "Apache 2.0"}
authors = [
    {name = "Alex Al-Saffar", email = "programmer@usa.com"},
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.13"
dependencies = [
    "cryptography>=44.0.2",
    "fire>=0.7.0",
    "joblib>=1.5.2",
    "paramiko>=3.5.1",
    "randomname>=0.2.1",
    "requests>=2.32.5",
    "rich>=14.0.0",
    "tenacity>=9.1.2",
    "psutil>=7.0.0",
    "gitpython>=3.1.44",
    "pyfzf>=0.3.1",
    "rclone-python>=0.1.23",
    "pyjson5>=1.6.9",
    "pyyaml>=6.0.2",
    "toml>=0.10.2",
    "tomli>=2.2.1",
]

[project.optional-dependencies]
windows = [
    "pywin32"
]
# docs extra for documentation generation
docs = [
    "pdoc>=15.0.2"
]

[project.urls]
"Homepage" = "https://github.com/thisismygitrepo/machineconfig"
"Bug Tracker" = "https://github.com/thisismygitrepo/machineconfig/issues"

[project.scripts]
devops = "machineconfig.scripts.python.devops:args_parser"
choose_wezterm_theme = "machineconfig.scripts.python.choose_wezterm_theme:main"
cloud_copy = "machineconfig.scripts.python.cloud_copy:arg_parser"
cloud_mount = "machineconfig.scripts.python.cloud_mount:main"
cloud_repo_sync = "machineconfig.scripts.python.cloud_repo_sync:args_parser"
cloud_sync = "machineconfig.scripts.python.cloud_sync:args_parser"
croshell = "machineconfig.scripts.python.croshell:build_parser"
fire = "machineconfig.scripts.python.fire_jobs:main"
fire_agents = "machineconfig.scripts.python.fire_agents:main"
ftpx = "machineconfig.scripts.python.ftpx:main"
mcinit = "machineconfig.scripts.python.ai.mcinit:main"
mount_nfs = "machineconfig.scripts.python.mount_nfs:main"
mount_nw_drive = "machineconfig.scripts.python.mount_nw_drive:main"
repos = "machineconfig.scripts.python.repos:main"
start_slidev = "machineconfig.scripts.python.start_slidev:main"
wifi_conn = "machineconfig.scripts.python.wifi_conn:main"
kill_process = "machineconfig.utils.procs:main"

[tool.setuptools]
package-dir = {"" = "src"}
license-files = []

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv.sources]
crocodile = { path = "../crocodile", editable = true }

[dependency-groups]
dev = [
    "cleanpy>=0.5.1",
    "ipdb>=0.13.13",
    "ipykernel>=6.30.1",
    "ipython>=9.5.0",
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "pudb>=2025.1",
    "pylint>=3.3.8",
    "pyrefly>=0.33.1",
    "pyright>=1.1.405",
    "ruff>=0.13.1",
    "ty>=0.0.1a21",
    "types-mysqlclient>=2.2.0.20250915",
    "types-paramiko>=4.0.0.20250822",
    "types-pytz>=2025.2.0.20250809",
    "types-pyyaml>=6.0.12.20250915",
    "types-requests>=2.32.4.20250913",
    "types-sqlalchemy>=1.4.53.38",
    "types-toml>=0.10.8.20240310",
    "types-urllib3>=1.26.25.14",
]
docs = [
    "pdoc>=15.0.2",
]
